Extract surface capabilities to own module
