+++ vulkano/src/swapchain/mod.rs
@@ -191,137 +191,33 @@
 //!
 
 use std::sync::atomic::AtomicBool;
-use vk;
 
-pub use self::surface::Capabilities;
+pub use self::capabilities::Capabilities;
+pub use self::capabilities::PresentMode;
+pub use self::capabilities::SupportedPresentModes;
+pub use self::capabilities::SupportedPresentModesIter;
+pub use self::capabilities::SurfaceTransform;
+pub use self::capabilities::CompositeAlpha;
+pub use self::capabilities::SupportedCompositeAlpha;
+pub use self::capabilities::SupportedCompositeAlphaIter;
+pub use self::capabilities::ColorSpace;
+pub use self::capabilities::SupportedSurfaceTransforms;
+pub use self::capabilities::SupportedSurfaceTransformsIter;
 pub use self::surface::Surface;
-pub use self::surface::PresentMode;
-pub use self::surface::SupportedPresentModes;
-pub use self::surface::SupportedPresentModesIter;
-pub use self::surface::SurfaceTransform;
-pub use self::surface::CompositeAlpha;
-pub use self::surface::SupportedCompositeAlpha;
-pub use self::surface::SupportedCompositeAlphaIter;
-pub use self::surface::ColorSpace;
 pub use self::surface::SurfaceCreationError;
 pub use self::swapchain::AcquireError;
 pub use self::swapchain::PresentFuture;
 pub use self::swapchain::Swapchain;
 pub use self::swapchain::SwapchainAcquireFuture;
 
+mod capabilities;
 pub mod display;
 mod surface;
 mod swapchain;
 
-/// List of supported composite alpha modes.
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub struct SupportedSurfaceTransforms {
-    pub identity: bool,
-    pub rotate90: bool,
-    pub rotate180: bool,
-    pub rotate270: bool,
-    pub horizontal_mirror: bool,
-    pub horizontal_mirror_rotate90: bool,
-    pub horizontal_mirror_rotate180: bool,
-    pub horizontal_mirror_rotate270: bool,
-    pub inherit: bool,
-}
-
-impl SupportedSurfaceTransforms {
-    /// Builds a `SupportedSurfaceTransforms` with all fields set to false.
-    #[inline]
-    pub fn none() -> SupportedSurfaceTransforms {
-        SupportedSurfaceTransforms {
-            identity: false,
-            rotate90: false,
-            rotate180: false,
-            rotate270: false,
-            horizontal_mirror: false,
-            horizontal_mirror_rotate90: false,
-            horizontal_mirror_rotate180: false,
-            horizontal_mirror_rotate270: false,
-            inherit: false,
-        }
-    }
-
-    #[inline]
-    fn from_bits(val: vk::SurfaceTransformFlagsKHR) -> SupportedSurfaceTransforms {
-        macro_rules! v {
-            ($val:expr, $out:ident, $e:expr, $f:ident) => (
-                if ($val & $e) != 0 { $out.$f = true; }
-            );
-        }
-
-        let mut result = SupportedSurfaceTransforms::none();
-        v!(val, result, vk::SURFACE_TRANSFORM_IDENTITY_BIT_KHR, identity);
-        v!(val, result, vk::SURFACE_TRANSFORM_ROTATE_90_BIT_KHR, rotate90);
-        v!(val, result, vk::SURFACE_TRANSFORM_ROTATE_180_BIT_KHR, rotate180);
-        v!(val, result, vk::SURFACE_TRANSFORM_ROTATE_270_BIT_KHR, rotate270);
-        v!(val, result, vk::SURFACE_TRANSFORM_HORIZONTAL_MIRROR_BIT_KHR, horizontal_mirror);
-        v!(val, result, vk::SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_90_BIT_KHR,
-                        horizontal_mirror_rotate90);
-        v!(val, result, vk::SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_180_BIT_KHR,
-                        horizontal_mirror_rotate180);
-        v!(val, result, vk::SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_270_BIT_KHR,
-                        horizontal_mirror_rotate270);
-        v!(val, result, vk::SURFACE_TRANSFORM_INHERIT_BIT_KHR, inherit);
-        result
-    }
-
-    /// Returns true if the given `SurfaceTransform` is in this list.
-    #[inline]
-    pub fn supports(&self, value: SurfaceTransform) -> bool {
-        match value {
-            SurfaceTransform::Identity => self.identity,
-            SurfaceTransform::Rotate90 => self.rotate90,
-            SurfaceTransform::Rotate180 => self.rotate180,
-            SurfaceTransform::Rotate270 => self.rotate270,
-            SurfaceTransform::HorizontalMirror => self.horizontal_mirror,
-            SurfaceTransform::HorizontalMirrorRotate90 => self.horizontal_mirror_rotate90,
-            SurfaceTransform::HorizontalMirrorRotate180 => self.horizontal_mirror_rotate180,
-            SurfaceTransform::HorizontalMirrorRotate270 => self.horizontal_mirror_rotate270,
-            SurfaceTransform::Inherit => self.inherit,
-        }
-    }
-
-    /// Returns an iterator to the list of supported composite alpha.
-    #[inline]
-    pub fn iter(&self) -> SupportedSurfaceTransformsIter {
-        SupportedSurfaceTransformsIter(self.clone())
-    }
-}
-
-/// Enumeration of the `SurfaceTransform` that are supported.
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub struct SupportedSurfaceTransformsIter(SupportedSurfaceTransforms);
-
-impl Iterator for SupportedSurfaceTransformsIter {
-    type Item = SurfaceTransform;
-
-    #[inline]
-    fn next(&mut self) -> Option<SurfaceTransform> {
-        if self.0.identity { self.0.identity = false; return Some(SurfaceTransform::Identity); }
-        if self.0.rotate90 { self.0.rotate90 = false; return Some(SurfaceTransform::Rotate90); }
-        if self.0.rotate180 { self.0.rotate180 = false; return Some(SurfaceTransform::Rotate180); }
-        if self.0.rotate270 { self.0.rotate270 = false; return Some(SurfaceTransform::Rotate270); }
-        if self.0.horizontal_mirror { self.0.horizontal_mirror = false; return Some(SurfaceTransform::HorizontalMirror); }
-        if self.0.horizontal_mirror_rotate90 { self.0.horizontal_mirror_rotate90 = false; return Some(SurfaceTransform::HorizontalMirrorRotate90); }
-        if self.0.horizontal_mirror_rotate180 { self.0.horizontal_mirror_rotate180 = false; return Some(SurfaceTransform::HorizontalMirrorRotate180); }
-        if self.0.horizontal_mirror_rotate270 { self.0.horizontal_mirror_rotate270 = false; return Some(SurfaceTransform::HorizontalMirrorRotate270); }
-        if self.0.inherit { self.0.inherit = false; return Some(SurfaceTransform::Inherit); }
-        None
-    }
-}
-
-impl Default for SurfaceTransform {
-    #[inline]
-    fn default() -> SurfaceTransform {
-        SurfaceTransform::Identity
-    }
-}
-
 /// Internal trait so that creating/destroying a swapchain can access the surface's "has_swapchain"
 /// flag.
+// TODO: use pub(crate) maybe?
 unsafe trait SurfaceSwapchainLock {
     fn flag(&self) -> &AtomicBool;
 }
